Split UI inbox code into store/data modules and remove dead inbox deadcode
