--- ui/src/inbox/index.ts
@@ -1,1 +1,0 @@
-export * from './inbox'
