--- ui/src/inbox/types.ts
@@ -1,7 +1,0 @@
-export type {
-  InboxConversationState,
-  InboxResponse,
-  ConversationInboxRefreshResponse,
-  ThreadInboxEntry,
-  ThreadInboxResponse,
-} from '../data/inbox'
