+++ ui/src/store/inbox.test.ts
@@ -3,10 +3,9 @@ import { describe, expect, it } from 'vitest'
 import {
   bootstrapInboxState,
   ensureInboxConversations,
-  mergeInboxNotificationRefresh,
 } from './inbox'
 import type { ChannelInfo } from '../components/channels/types'
-import type { InboxConversationState } from './types'
+import type { InboxConversationState } from '../data/inbox'
 
 function makeConversation(
   overrides: Partial<InboxConversationState> = {}
@@ -36,74 +35,6 @@ function makeChannel(overrides: Partial<ChannelInfo> = {}): ChannelInfo {
     ...overrides,
   }
 }
-
-describe('mergeInboxNotificationRefresh', () => {
-  it('applies server conversation + thread rows from inbox-notification', () => {
-    const state = bootstrapInboxState([makeConversation({ latestSeq: 2, unreadCount: 0 })])
-
-    const next = mergeInboxNotificationRefresh(state, {
-      conversation: {
-        conversationId: 'conversation-1',
-        conversationName: 'general',
-        conversationType: 'channel',
-        latestSeq: 5,
-        lastReadSeq: 2,
-        unreadCount: 2,
-        threadUnreadCount: 0,
-        lastMessageId: 'm-new',
-        lastMessageAt: '2026-03-30T01:00:00Z',
-      },
-      thread: {
-        conversationId: 'conversation-1',
-        threadParentId: 'parent-1',
-        latestSeq: 4,
-        lastReadSeq: 1,
-        unreadCount: 1,
-        lastReplyMessageId: 'reply-1',
-        lastReplyAt: '2026-03-30T00:30:00Z',
-      },
-    })
-
-    expect(next.conversations['conversation-1']).toEqual(
-      expect.objectContaining({
-        latestSeq: 5,
-        lastReadSeq: 2,
-        unreadCount: 2,
-        lastMessageId: 'm-new',
-      })
-    )
-    expect(next.threads['conversation-1:parent-1']).toEqual(
-      expect.objectContaining({
-        unreadCount: 1,
-        latestSeq: 4,
-        lastReadSeq: 1,
-      })
-    )
-  })
-
-  it('ignores stale payloads when latestSeq regressed', () => {
-    const state = bootstrapInboxState([
-      makeConversation({ latestSeq: 10, unreadCount: 1 }),
-    ])
-
-    const next = mergeInboxNotificationRefresh(state, {
-      conversation: {
-        conversationId: 'conversation-1',
-        conversationName: 'general',
-        conversationType: 'channel',
-        latestSeq: 9,
-        lastReadSeq: 2,
-        unreadCount: 99,
-        threadUnreadCount: 0,
-        lastMessageId: 'old',
-        lastMessageAt: null,
-      },
-    })
-
-    expect(next.conversations['conversation-1']?.latestSeq).toBe(10)
-    expect(next.conversations['conversation-1']?.unreadCount).toBe(1)
-  })
-})
 
 describe('bootstrapInboxState', () => {
   it('seeds zeroed entries for joined conversations missing from the inbox snapshot', () => {
@@ -163,5 +94,3 @@ describe('ensureInboxConversations', () => {
     )
   })
 })
-
-
