+++ ui/src/types/index.ts
@@ -46,7 +46,6 @@ export type {
 export type {
   InboxConversationState,
   InboxResponse,
-  ConversationInboxRefreshResponse,
   ThreadInboxEntry,
   ThreadInboxResponse,
 } from '../data/inbox'
